Split test_basic
